backend/ugc_service/tests/requirements.txt
pytest==7.2.1
pytest-asyncio==0.20.3
aiohttp==3.8.4
pydantic==1.10.5
protobuf==4.21.12
grpcio-tools==1.51.1
multidict==6.0.4

pytest-postgresql==4.1.0
psycopg2-binary==2.9.5
psycopg[binary]==3.1.8
